Put logic in files.go; use existing MoveFile()
