+++ internal/utils/files.go
@@ -4,6 +4,9 @@ import (
 	"fmt"
 	"io"
 	"os"
+
+	"github.com/doron-cohen/antidot/internal/tui"
+	"github.com/otiai10/copy"
 )
 
 func FileExists(filename string) bool {
@@ -38,4 +41,33 @@ func MoveFile(sourcePath, destPath string) error {
 		return fmt.Errorf("Failed removing original file: %s", err)
 	}
 	return nil
+}
+
+func MoveDirectory(source, dest string) error {
+	err := copy.Copy(source, dest)
+	if err != nil {
+		return err
+	}
+
+	err = os.RemoveAll(source)
+	if err != nil {
+		tui.Warn("Failed to remove original directory: %s", err)
+	}
+
+	return nil
+}
+
+// Try to move file/directory with os.Rename and if that fails, do a copy + delete
+func MovePath(source, dest string) error {
+	err := os.Rename(source, dest)
+	if err == nil {
+		return nil
+	}
+
+	fi, err := os.Stat(source)
+	if fi.Mode().IsDir() {
+		return MoveDirectory(source, dest)
+	}
+
+	return MoveFile(source, dest)
 }
